feat: allow raw parquet upload

+++ src/commands/dataset/convert.rs
@@ -448,7 +448,7 @@ pub async fn convert(args: Convert, global: GlobalArgs) -> Result<()> {
     pb.set_length(file_len);
     pb.set_position(0);
 
-    let stream = stream.inspect_ok(|item| pb.set_position(item.end as u64));
+    let stream = stream.inspect_ok(|item| pb.set_position(item.end));
 
     let stream = read::from_value_stream(stream, schema.clone(), read_options).map_err(|err| {
         error::user(
